requirements.txt
pocketflow
google-generativeai  #pip install google-genai==1.25.0